header component refactor
