--- src/utils/pickRandomWinner.js
@@ -1,5 +1,0 @@
-const pickRandomWinner = (collectionData) => {
-  return collectionData[Math.floor(Math.random() * collectionData.length)];
-};
-
-export default pickRandomWinner;
